+++ packages/components/src/FilterBar/index.ts
@@ -1,0 +1,4 @@
+export * from "./FilterBar"
+export * from "./context/FilterBarContext"
+export * from "./context/types"
+export * from "./types"