+++ packages/components/src/Workflow/_docs/Workflow.stories.tsx
@@ -1,19 +1,20 @@
 import React from "react"
 import { Meta, StoryFn, StoryObj } from "@storybook/react"
-import { Button } from "@kaizen/button"
-import VisibleIcon from "@kaizen/component-library/icons/visible.icon.svg"
+import { Button } from "~components/Button"
+import { CloseIcon } from "~icons/CloseIcon"
+import { VisibleIcon } from "~icons/VisibleIcon"
 import { Workflow, WorkflowProps } from "../"
 import { WorkflowControls } from "./controls"
 
 const MockContent = (): JSX.Element => (
   <div className="flex flex-col flex-1 m-24 items-center">
     <div className="max-w-[1392px] w-100">
-      <h3>Name and schedule the self-reflection cycle</h3>
+      <h2>Name and schedule the self-reflection cycle</h2>
       <form
         id="workflow-form-id"
         className="bg-white rounded-default shadow-sm my-32 p-64"
       >
-        <h4>Name the cycle</h4>
+        <h3>Name the cycle</h3>
         <label
           className="block font-weight-paragraph-bold"
           htmlFor="input-workflow"
@@ -45,14 +46,12 @@ const defaultArgs = {
   previousAction: <Button reversed label="Back" />,
   nextAction: <Button reversed label="Next" />,
   headerActions: [
-    <Workflow.WorkflowExit
-      key="would-use-uui"
-      exitLabel="Save and close"
-      exitTitle="Before you exit"
-      exitDescription="Your content has not yet been saved. Click the button below or discard the changes"
-      confirmExitLabel="Close and save"
-      dismissExitLabel="Dismiss"
-      onExit={(): void => alert("mock example of a save action")}
+    <Button
+      key="would-use-uui-2"
+      label="Save and close"
+      icon={<CloseIcon />}
+      secondary
+      iconPosition="end"
     />,
   ],
   children: <MockContent />,
@@ -74,6 +73,7 @@ export const Playground: StoryObj<typeof meta> = {
         sourceState: "shown",
       },
     },
+    chromatic: { disable: false },
   },
 }
 
@@ -94,18 +94,17 @@ export const MultipleActions: StoryObj<typeof meta> = {
       <Button
         key="would-use-uui-1"
         label="Preview"
-        icon={VisibleIcon}
+        icon={<VisibleIcon />}
         secondary
         iconPosition="start"
       />,
-      <Workflow.WorkflowExit
+      <Button
         key="would-use-uui-2"
-        exitLabel="Save and close"
-        exitTitle="Before you exit"
-        exitDescription="Your content has not yet been saved. Click the button below or discard the changes"
-        confirmExitLabel="Close and save"
-        dismissExitLabel="Dismiss"
-        onExit={(): void => alert("mock example of a save action")}
+        label="Save and close"
+        icon={<CloseIcon />}
+        secondary
+        iconPosition="end"
+        onClick={(): void => alert("mock example of a save action")}
       />,
     ],
   },
@@ -114,7 +113,7 @@ export const MultipleActions: StoryObj<typeof meta> = {
 export const FinalStep: StoryObj<typeof meta> = {
   args: {
     workflowName: "Create a self-reflection cycle",
-    stepName: "Settings",
+    stepName: "Schedule",
     steps: ["Settings", "Questions", "Preview", "Employees", "Schedule"],
     isComplete: false,
     status: {
@@ -123,14 +122,13 @@ export const FinalStep: StoryObj<typeof meta> = {
     },
     nextAction: <Button reversed label="Next" />,
     headerActions: [
-      <Workflow.WorkflowExit
-        key="would-use-uui"
-        exitLabel="Save and close"
-        exitTitle="Before you exit"
-        exitDescription="Your content has not yet been saved. Click the button below or discard the changes"
-        confirmExitLabel="Close and save"
-        dismissExitLabel="Dismiss"
-        onExit={(): void => alert("mock example of a save action")}
+      <Button
+        key="would-use-uui-2"
+        label="Save and close"
+        icon={<CloseIcon />}
+        secondary
+        iconPosition="end"
+        onClick={(): void => alert("mock example of a save action")}
       />,
     ],
   },
@@ -148,16 +146,18 @@ export const CompletedWorkflow: StoryObj<typeof meta> = {
     },
     nextAction: <Button reversed label="Next" />,
     headerActions: [
-      <Workflow.WorkflowExit
-        key="would-use-uui"
-        exitLabel="Save and close"
-        exitTitle="Before you exit"
-        exitDescription="Your content has not yet been saved. Click the button below or discard the changes"
-        confirmExitLabel="Close and save"
-        dismissExitLabel="Dismiss"
-        onExit={(): void => alert("mock example of a save action")}
+      <Button
+        key="would-use-uui-2"
+        label="Save and close"
+        icon={<CloseIcon />}
+        secondary
+        iconPosition="end"
+        onClick={(): void => alert("mock example of a save action")}
       />,
     ],
+  },
+  parameters: {
+    chromatic: { disable: false },
   },
 }
 
@@ -172,7 +172,7 @@ export const ComposableWorkflow: StoryFn<WorkflowProps> = ({
   nextAction,
   ...restProps
 }) => (
-  <div {...restProps}>
+  <Workflow.Wrapper {...restProps}>
     <Workflow.Header
       workflowName={workflowName}
       stepName={stepName}
@@ -195,5 +195,38 @@ export const ComposableWorkflow: StoryFn<WorkflowProps> = ({
       nextAction={nextAction}
       previousAction={previousAction}
     />
-  </div>
+  </Workflow.Wrapper>
 )
+
+ComposableWorkflow.parameters = {
+  chromatic: { disable: false },
+}
+
+export const ResponsiveWorkflow: StoryObj<typeof meta> = {
+  parameters: {
+    docs: {
+      canvas: {
+        sourceState: "hidden",
+      },
+    },
+    viewport: {
+      viewports: {
+        vieportZoomed: {
+          name: "Simulate 400% zoom",
+          styles: {
+            width: "500px",
+            height: "800px",
+          },
+          type: "mobile",
+        },
+      },
+      defaultViewport: "vieportZoomed",
+    },
+    chromatic: {
+      disable: false,
+      viewports: [500, 1200],
+    },
+  },
+}
+
+ResponsiveWorkflow.args = { ...defaultArgs }