+++ packages/components/src/Workflow/subcomponents/Main/Main.tsx
@@ -1,18 +1,20 @@
-import React, { PropsWithChildren, HTMLAttributes } from "react"
+import React, { HTMLAttributes, ReactNode } from "react"
 import classnames from "classnames"
 import { OverrideClassName } from "~types/OverrideClassName"
+import styles from "./Main.module.scss"
 
-export type WorkflowMainProps = PropsWithChildren<
-  OverrideClassName<HTMLAttributes<HTMLDivElement>>
->
+export type WorkflowMainProps = OverrideClassName<
+  HTMLAttributes<HTMLDivElement>
+> & {
+  children: ReactNode
+}
 
-/** A wrapper for the exit trigger (and other) actions of the Header Workflow  */
 export const Main = ({
   children,
   classNameOverride,
   ...restProps
 }: WorkflowMainProps): JSX.Element => (
-  <main className={classnames([classNameOverride])} {...restProps}>
+  <main className={classnames([styles.main, classNameOverride])} {...restProps}>
     {children}
   </main>
 )