+++ packages/components/src/Workflow/subcomponents/index.tsx
@@ -1,3 +1,0 @@
-export * from "./Footer"
-export * from "./Header"
-export * from "./Main"