+++ packages/date-picker/docs/FloatingCalendarWrapper.stories.tsx
@@ -11,7 +11,8 @@ const SUPPORTED_LOCALES = ["en-US", "en-AU"]
 
 export default {
   tags: ["autodocs"],
-  title: "Components/Date Picker/Subcomponents/Floating Calendar Wrapper",
+  title:
+    "Components/Date Selection/Date Picker/Subcomponents/Floating Calendar Wrapper",
   component: FloatingCalendarWrapper,
   parameters: {
     docs: {