+++ packages/notification/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 1.5.10
+
+### Patch Changes
+
+- [#3731](https://github.com/cultureamp/kaizen-design-system/pull/3731) [`3421da907`](https://github.com/cultureamp/kaizen-design-system/commit/3421da907e3e4e235da4f6a733e33cffbd258a74) - InlineNotification width has changed from `100%` to `auto` so that it can be used side by side with other components. Should not affect any current usages.
+
+## 1.5.9
+
+### Patch Changes
+
+- [#3677](https://github.com/cultureamp/kaizen-design-system/pull/3677) [`06c48bb35`](https://github.com/cultureamp/kaizen-design-system/commit/06c48bb3593cd348c393369b6bc17d04a997c9e0) - Bump outdated `component-library` and `draft-form` dependencies to latest
+
 ## 1.5.8
 
 ### Patch Changes